+++ src/AwsV3BridgeServiceProvider.php
@@ -3,8 +3,9 @@
 namespace MvbCoding\Silex;
 
 use Aws\SimpleDb\SimpleDbClient;
+use Pimple\Container;
+use Pimple\ServiceProviderInterface;
 use Silex\Application;
-use Silex\ServiceProviderInterface;
 
 /**
  * AWS SDK for PHP service provider for Silex applications
@@ -13,19 +14,15 @@ class AwsV3BridgeServiceProvider implements ServiceProviderInterface
 {
     const VERSION = '3.0.0';
 
-    public function register(Application $app)
+    public function register(Container $container)
     {
-        $app['aws.simpledb'] = $app->share(function (Application $app) {
-            $config = isset($app['aws.config']) ? $app['aws.config'] : [];
+        $container['aws.simpledb'] = function ($container) {
+            $config = isset($container['aws.config']) ? $container['aws.config'] : [];
 
             return new SimpleDbClient($config + ['ua_append' => [
                 'Silex/' . Application::VERSION,
                 'SXMOD/' . self::VERSION,
             ]]);
-        });
-    }
-
-    public function boot(Application $app)
-    {
+        };
     }
 }